requirements.txt
flask>=3.0
line-bot-sdk>=3.0
groq>=0.4
deep-translator>=1.11
edge-tts>=6.1
aiohttp>=3.9
python-dotenv>=1.0
opencc-python-reimplemented>=0.1.7
